AnimFormat : Fix keyframes with "In"/"Out" values+tangents
Previously keyframes were being outputted with the same "time" value, effectively overwriting the "In" tangent with the following "Out" tangent when imported into Maya etc. This fix properly assigns in and out tangents to keyframes that need them, and keyframes with unique Out values are placed just slightly later to prevent overwriting of the previous frame.