CLR0s With UserData No Longer Crash on PostProcess
Previously, attempting to save a CLR0 that had UserData entries defined would crash the program.
This happened because the block that does the PostProcess calls for the CLR0MaterialNodes was changing the value of the dataAddress parameter before we reached the UserEntries PostProcess call; so the _userEntries.PostProcess() call was being done with a bad address parameter.
Fixed by giving the CLR0MaterialNode block a local variable for holding the addresses it uses so it doesn't overwrite the original parameter.