Fix collada dae import for geometry with shared input offsets
The Collada dae importer was not accounting for geometries that used inputs with shared offset values and assumed that there are always as many indices per facePoint as there were inputs.

This was causing the decoder to read/write vertices and indices to/from the wrong memory addresses, which causes undefined behavior.

This should help prevent BC from crashing when importing some dae files that would otherwise work in other tools.